implemented a Timer and Health bar system in full
Also performed slight refactors/reduction in
certain areas of the code to aim to reduce amount
of code on the page.

Will be working to reduce indentation to clean the look
of the code in JS, and move inline CSS attributes into a
seperate CSS page in order to keep to clean coding standards outlined
for HTML5 and CSS3 standards.