editing back on github as opposed to VScode.
managed to finish the majority of my work using a youtube explanation given online on building a javascript game.

I am now reorganising the sprites to have the correct file directories within my filepaths and then will work on the responsiveness of the page for a user using a mobile advice.